[es6-bookreader-core/Navbar] create Navbar.js
